StockBot/requirements.txt
fastapi
uvicorn
jinja2
python-dotenv
anthropic
requests
aiohttp